+++ app/routes/board._index/types/PasswordCheckOptionType.ts
@@ -1,0 +1,10 @@
+
+export const PWCheckOption = {
+  ViewDetail: "viewDetail",
+  Delete: "delete",
+  Edit: "edit",
+  None: "none",
+} as const;
+
+export type PWCheckOptionType =
+  (typeof PWCheckOption)[keyof typeof PWCheckOption];
